reqs/requirements.txt
langchain==0.3.26
langchain_openai==0.3.28
python-dotenv==1.1.1
